feat(ehr): moving ehr internal too
Ref: #1040
