+++ packages/api/src/routes/index.ts
@@ -4,9 +4,9 @@ import biometrics from "./biometrics";
 import body from "./body";
 import connect from "./connect";
 import ehr from "./ehr";
-import ehrInternal from "./ehr-internal";
 import feedback from "./feedback";
 import { reportClientErrors } from "./helpers/report-client-errors";
+import ehrInternal from "./internal/ehr";
 import internal from "./internal/internal";
 import jwtToken from "./internal/jwt-token/ehr";
 import medical from "./medical";
